delete user portrait on error or replacement, organize middleware
